back/python/tag_manager/requirements.txt
fastapi~=0.111.0
python-dotenv~=1.0.1
uvicorn~=0.29.0
pymongo~=4.7.0